requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
speechbrain>=0.5.16
pydub>=0.25.1
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
